% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{accumulate_hero_error}
\alias{accumulate_hero_error}
\title{Accumulate a heRo_error (INTERNAL).}
\usage{
accumulate_hero_error(error_obj, context_msg)
}
\arguments{
\item{error_obj}{An object of class `heRo_error`.}

\item{context_msg}{A character string describing the context.}
}
\value{
Invisibly returns NULL.
}
\description{
Adds a heRo_error object to the internal error list.
}
