% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{accumulate_oq_error}
\alias{accumulate_oq_error}
\title{Accumulate an oq_error (INTERNAL).}
\usage{
accumulate_oq_error(error_obj, context_msg)
}
\arguments{
\item{error_obj}{An object of class `oq_error`.}

\item{context_msg}{A character string describing the context.}
}
\value{
Invisibly returns NULL.
}
\description{
Adds an oq_error object to the internal error list.
}
