% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{add_psm_transition}
\alias{add_psm_transition}
\title{Add PSM Transitions to Model}
\usage{
add_psm_transition(model, endpoint, time_unit, formula)
}
\arguments{
\item{model}{A oq_model_builder object}

\item{endpoint}{Character string for the endpoint}

\item{time_unit}{Character string for the time unit}

\item{formula}{An unquoted R expression for the transition}
}
\value{
The modified model object
}
\description{
Add transitions for a PSM (Partitioned Survival Model).
}
