% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_output.R
\name{add_time_columns}
\alias{add_time_columns}
\title{Add Time Unit Columns to Transition Data}
\usage{
add_time_columns(df, time_unit, results)
}
\arguments{
\item{df}{Data frame with a cycle column}

\item{time_unit}{Requested time unit}

\item{results}{Full results object for settings access}
}
\value{
Data frame with time columns added
}
\description{
Add Time Unit Columns to Transition Data
}
\keyword{internal}
