% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{add_tree_node}
\alias{add_tree_node}
\title{Add a Tree Node}
\usage{
add_tree_node(model, tree_name, node, parent = NA, formula, tags = NA)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{tree_name}{Character string for the tree name}

\item{node}{Character string for the node name}

\item{parent}{Character string for the parent node name (NA for root)}

\item{formula}{An unquoted R expression for the node probability}

\item{tags}{Optional character string for node tags}
}
\value{
The modified model object
}
\description{
Add a node to a decision tree in the model. Creates the trees tibble if it
doesn't exist yet.
}
