% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{aggregate_trace}
\alias{aggregate_trace}
\title{Aggregate Trace Results}
\usage{
aggregate_trace(segments)
}
\arguments{
\item{segments}{Data frame of segments for a single strategy with normalized weights}
}
\value{
A matrix containing the weighted average trace
}
\description{
Calculates weighted average of trace results across segments.
}
\keyword{internal}
