% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{aggregate_values}
\alias{aggregate_values}
\title{Aggregate Values Results}
\usage{
aggregate_values(segments)
}
\arguments{
\item{segments}{Data frame of segments for a single strategy with normalized weights}
}
\value{
A matrix containing the weighted average values
}
\description{
Calculates weighted average of values results across segments.
}
\keyword{internal}
