% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsa_vbp.R
\name{analyze_twsa_vbp_results}
\alias{analyze_twsa_vbp_results}
\title{Analyze TWSA VBP Results}
\usage{
analyze_twsa_vbp_results(segments, aggregated, vbp_spec, twsa_metadata, model)
}
\arguments{
\item{segments}{Segment results}

\item{aggregated}{Aggregated results}

\item{vbp_spec}{VBP specification}

\item{twsa_metadata}{TWSA metadata tibble}

\item{model}{Parsed model object}
}
\value{
Tibble with VBP equations per run_id
}
\description{
Calculates VBP equations for each TWSA run.
}
\keyword{internal}
