% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_codegen.R
\name{append_section}
\alias{append_section}
\title{Append a code section with a blank line separator, only if non-empty}
\usage{
append_section(code, section_code)
}
\description{
Append a code section with a blank line separator, only if non-empty
}
\keyword{internal}
