% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{as.character.oq_formula}
\alias{as.character.oq_formula}
\title{Convert a Formula to Character}
\usage{
\method{as.character}{oq_formula}(x, ...)
}
\arguments{
\item{x}{An oq_formula object.}

\item{...}{additional unused arguments.}
}
\value{
An atomic character vector representing the R-expression.
}
\description{
Takes an oq_formula object and converts it to its character representation.
}
