% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.r
\name{as.lf_markov_trans}
\alias{as.lf_markov_trans}
\alias{as.lf_markov_trans.lf_markov_trans}
\alias{as.lf_markov_trans.data.frame}
\title{Coerce to Long-Format Markov Transitions}
\usage{
as.lf_markov_trans(x)

\method{as.lf_markov_trans}{lf_markov_trans}(x)

\method{as.lf_markov_trans}{data.frame}(x)
}
\arguments{
\item{x}{An object to coerce.}
}
\value{
An object of class `lf_markov_trans`.
}
\description{
S3 generic function that coerces objects to the `lf_markov_trans` class,
which represents Markov transition definitions in long format.
}
