% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{as.oq_formula}
\alias{as.oq_formula}
\title{Convert to Formula}
\usage{
as.oq_formula(x)
}
\arguments{
\item{x}{Object to be converted to oq_formula.}
}
\value{
An oq_formula object.
}
\description{
Convert an object to an oq_formula object.
}
\keyword{internal}
