% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{as_json}
\alias{as_json}
\title{Convert Model to JSON String}
\usage{
as_json(model)
}
\arguments{
\item{model}{A openqaly model object}
}
\value{
A JSON string
}
\description{
Convert a openqaly model object to a JSON string representation.
}
