% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{as_r_code}
\alias{as_r_code}
\title{Convert Model to R Code String}
\usage{
as_r_code(model)
}
\arguments{
\item{model}{A heRomodel object}
}
\value{
A character vector of R code lines
}
\description{
Convert a heRomodel object to executable R code.
}
