% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_distributions.R
\name{beta}
\alias{beta}
\title{Beta distribution}
\usage{
beta(mean = NULL, sd = NULL, shape1 = NULL, shape2 = NULL)
}
\arguments{
\item{mean}{Mean of parameter (alternative parameterization)}

\item{sd}{Standard deviation of parameter (alternative parameterization)}

\item{shape1}{Alpha parameter (traditional parameterization)}

\item{shape2}{Beta parameter (traditional parameterization)}
}
\description{
The `beta` function can be used to sample a beta-distributed
parameter, commonly used for probabilities bounded between 0 and 1.
Can be specified using either (mean, sd) or (shape1, shape2) parameterization.
}
