% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openqaly.R
\docType{data}
\name{C}
\alias{C}
\title{Complementary Probability Sentinel}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
C
}
\description{
A sentinel value used in decision tree probability calculations to indicate
that a node's probability should be computed as the complement of the sum
of its sibling probabilities (i.e., 1 minus the sum of other probabilities
at the same level).
}
\keyword{datasets}
