% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.r
\name{calc_compl_probs}
\alias{calc_compl_probs}
\title{Calculate complementary probabilities in an evaluated transition matrix}
\usage{
calc_compl_probs(mat)
}
\arguments{
\item{mat}{A transition matrix array}
}
\value{
The matrix with complementary probabilities calculated
}
\description{
Calculate complementary probabilities in an evaluated transition matrix
}
\keyword{internal}
