% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm.R
\name{calculate_psm_custom_values}
\alias{calculate_psm_custom_values}
\title{Calculate Custom PSM Values}
\usage{
calculate_psm_custom_values(
  uneval_values,
  namespace,
  trace,
  value_names,
  state_names,
  n_cycles,
  half_cycle_method = "start"
)
}
\arguments{
\item{uneval_values}{Unevaluated values specifications}

\item{namespace}{Evaluated variables namespace}

\item{trace}{State occupancy trace matrix}

\item{value_names}{Vector of value names}

\item{state_names}{Vector of state names}

\item{n_cycles}{Number of cycles}

\item{half_cycle_method}{Half-cycle correction method}
}
\value{
A matrix of calculated values
}
\description{
Calculates residency and model-level values for Custom PSM models.
Transitional values are not supported.
}
\keyword{internal}
