% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm.R
\name{calculate_psm_trace_and_values}
\alias{calculate_psm_trace_and_values}
\title{Calculate PSM Trace and Values}
\usage{
calculate_psm_trace_and_values(
  survival_distributions,
  uneval_values,
  namespace,
  value_names,
  state_names,
  n_cycles,
  half_cycle_method = "start"
)
}
\arguments{
\item{survival_distributions}{List containing pfs and os distributions}

\item{uneval_values}{Unevaluated values specifications}

\item{namespace}{Evaluated variables namespace}

\item{value_names}{Vector of value names}

\item{state_names}{Vector of state names (should be 3 states)}

\item{n_cycles}{Number of cycles to simulate}
}
\value{
A list containing trace matrix and values matrix
}
\description{
Calculates the trace (state occupancy over time) and values for a partitioned survival model.
Uses PFS and OS distributions to determine state probabilities.
}
\keyword{internal}
