% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_ce_utils.R
\name{classify_base_case}
\alias{classify_base_case}
\title{Classify Base Case ICER}
\usage{
classify_base_case(icer_value)
}
\arguments{
\item{icer_value}{Numeric ICER value}
}
\value{
Character classification
}
\description{
Classify Base Case ICER
}
\keyword{internal}
