% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{convert_hero_formulas}
\alias{convert_hero_formulas}
\title{Convert heRoFormula Objects to Strings}
\usage{
convert_hero_formulas(obj)
}
\arguments{
\item{obj}{Object to convert}
}
\value{
Object with heRoFormula objects converted to strings
}
\description{
Recursively converts heRoFormula objects to character strings
throughout a nested data structure.
}
\keyword{internal}
