% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{convert_oq_formulas}
\alias{convert_oq_formulas}
\title{Convert oq_formula Objects to Strings}
\usage{
convert_oq_formulas(obj)
}
\arguments{
\item{obj}{Object to convert}
}
\value{
Object with oq_formula objects converted to strings
}
\description{
Recursively converts oq_formula objects to character strings
throughout a nested data structure.
}
\keyword{internal}
