% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert_settings_from_json}
\alias{convert_settings_from_json}
\title{Convert settings from JSON format}
\usage{
convert_settings_from_json(settings_df)
}
\arguments{
\item{settings_df}{A data frame containing settings from JSON}
}
\value{
A named list of settings
}
\description{
Convert settings from JSON format
}
