% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{define_namespace}
\alias{define_namespace}
\title{Define a Namespace Object}
\usage{
define_namespace(env, df, additional = NULL, ...)
}
\arguments{
\item{env}{An environment of pre-existing values}

\item{df}{A data frame of pre-existing parameter values}
}
\description{
A namespace object stores evaluated parameters by combining
a data frame for vector parameters with an environment
for non-vector parameters.
}
