% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{detect_model_format}
\alias{detect_model_format}
\title{Detect Model Format}
\usage{
detect_model_format(input)
}
\arguments{
\item{input}{The input to check}
}
\value{
A string indicating the format
}
\description{
Internal function to detect the format of a model input.
}
\keyword{internal}
