% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_plot.R
\name{.plot_eval_decision_tree}
\alias{.plot_eval_decision_tree}
\title{Plot an evaluated decision tree object}
\usage{
.plot_eval_decision_tree(x, locale = NULL)
}
\arguments{
\item{x}{An eval_decision_tree object}

\item{locale}{Locale for formatting, or NULL for US default}
}
\value{
A ggplot2/ggraph object
}
\description{
Plot an evaluated decision tree object
}
\keyword{internal}
