% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_vbp.R
\name{dsa_vbp}
\alias{dsa_vbp}
\title{DSA with Value-Based Pricing Add-on}
\description{
Functions for combining Deterministic Sensitivity Analysis (DSA) with
Value-Based Pricing (VBP) analysis. For each DSA parameter variation,
VBP sub-simulations are run to calculate the VBP relationship at each
parameter variation.
}
