% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{edit_custom_psm_transition}
\alias{edit_custom_psm_transition}
\title{Edit a Custom PSM Transition}
\usage{
edit_custom_psm_transition(model, state, formula)
}
\arguments{
\item{model}{An oq_model_builder object (must be type "custom_psm")}

\item{state}{Character string for the state name}

\item{formula}{An unquoted R expression for the new state probability}
}
\value{
The modified model object
}
\description{
Edit the formula of an existing Custom PSM state probability.
}
