% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{edit_decision_tree}
\alias{edit_decision_tree}
\title{Edit Decision Tree Configuration}
\usage{
edit_decision_tree(model, tree_name, duration, duration_unit)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{tree_name}{Character string referencing a tree in model$trees}

\item{duration}{An unquoted R expression for the duration}

\item{duration_unit}{Character string for the time unit}
}
\value{
The modified model object
}
\description{
Edit the decision tree configuration on the model.
}
