% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{edit_override_category}
\alias{edit_override_category}
\title{Edit an Override Category}
\usage{
edit_override_category(model, name, new_name, general)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{name}{Character string identifying the category to edit}

\item{new_name}{Optional new name for the category}

\item{general}{Optional logical for the general flag}
}
\value{
The modified model object
}
\description{
Modify the name or general flag of an existing override category.
}
