% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{edit_psm_transition}
\alias{edit_psm_transition}
\title{Edit a PSM Transition}
\usage{
edit_psm_transition(model, endpoint, formula, time_unit)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{endpoint}{Character string for the endpoint}

\item{formula}{An unquoted R expression for the new transition}

\item{time_unit}{Character string for the new time unit}
}
\value{
The modified model object
}
\description{
Edit an existing PSM transition's formula and/or time unit.
}
