% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{edit_transition}
\alias{edit_transition}
\title{Edit a Markov Transition}
\usage{
edit_transition(model, from_state, to_state, formula)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{from_state}{Character string specifying the source state}

\item{to_state}{Character string specifying the destination state}

\item{formula}{An unquoted R expression for the new transition probability}
}
\value{
The modified model object
}
\description{
Edit the formula of an existing Markov transition.
}
