% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{edit_tree_node}
\alias{edit_tree_node}
\title{Edit a Decision Tree Node}
\usage{
edit_tree_node(
  model,
  tree_name,
  node,
  formula,
  tags,
  parent,
  new_node_name,
  new_tree_name
)
}
\arguments{
\item{model}{An oq_model object}

\item{tree_name}{Character string for the tree name}

\item{node}{Character string for the node name}

\item{formula}{An unquoted R expression for the node probability}

\item{tags}{Optional character string for node tags}

\item{parent}{Character string for the parent node name}

\item{new_node_name}{Character string to rename the node}

\item{new_tree_name}{Character string to rename the entire tree}
}
\value{
The modified model object
}
\description{
Edit an existing tree node identified by tree_name and node.
}
