% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_codegen.R
\name{escape_quotes}
\alias{escape_quotes}
\title{Escape double quotes in a string for code generation}
\usage{
escape_quotes(s)
}
\description{
Escape double quotes in a string for code generation
}
\keyword{internal}
