% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsa.R
\name{estimate_twsa_param_size}
\alias{estimate_twsa_param_size}
\title{Estimate TWSA Parameter Grid Size}
\usage{
estimate_twsa_param_size(param)
}
\arguments{
\item{param}{Parameter specification from TWSA}
}
\value{
Integer estimate of value count
}
\description{
Returns an upper-bound estimate of the number of values
for one TWSA parameter dimension. Used to compute progress
totals before the actual build begins.
}
\keyword{internal}
