% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{export_model_json}
\alias{export_model_json}
\title{Export Model Results to JSON File}
\usage{
export_model_json(model_results, file, pretty = TRUE, auto_unbox = TRUE)
}
\arguments{
\item{model_results}{Results from run_model()}

\item{file}{Path to output JSON file}

\item{pretty}{Whether to format JSON with indentation (default TRUE)}

\item{auto_unbox}{Whether to automatically unbox single-element vectors (default TRUE)}
}
\description{
Prepares model results for JSON serialization and writes them to a JSON file.
This function first converts the results to a serializable format
using prepare_model_results_for_json(), then exports to JSON.
}
