% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_output.R
\name{export_trace_excel}
\alias{export_trace_excel}
\title{Export Trace to Excel}
\usage{
export_trace_excel(results, file, collapsed, separate_sheets, ...)
}
\description{
Export Trace to Excel
}
\keyword{internal}
