% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{extract_threshold_nmb}
\alias{extract_threshold_nmb}
\title{Extract NMB/CE Threshold Output}
\usage{
extract_threshold_nmb(aggregated, cond, summary_col, parsed_model)
}
\description{
Extract NMB/CE Threshold Output
}
\keyword{internal}
