% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{extract_threshold_outcomes_costs}
\alias{extract_threshold_outcomes_costs}
\title{Extract Outcomes/Costs Threshold Output}
\usage{
extract_threshold_outcomes_costs(aggregated, cond, summary_col)
}
\description{
Extract Outcomes/Costs Threshold Output
}
\keyword{internal}
