% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_output.R
\name{extract_trace_long}
\alias{extract_trace_long}
\title{Extract Trace as Long Format}
\usage{
extract_trace_long(
  source_data,
  states = NULL,
  cycles = NULL,
  time_unit = "cycle",
  trace_column = "collapsed_trace",
  metadata = NULL,
  strategy_name_field = "display_name",
  group_name_field = "display_name",
  state_name_field = "display_name"
)
}
\description{
Extract Trace as Long Format
}
\keyword{internal}
