% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_output.R
\name{extract_trace_wide}
\alias{extract_trace_wide}
\title{Extract Trace as Wide Format}
\usage{
extract_trace_wide(
  source_data,
  states = NULL,
  cycles = NULL,
  time_unit = "cycle",
  trace_column = "collapsed_trace",
  metadata = NULL,
  name_field = "display_name"
)
}
\description{
Extract Trace as Wide Format
}
\keyword{internal}
