% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_output.R
\name{extract_values_long}
\alias{extract_values_long}
\title{Extract Values as Long Format}
\usage{
extract_values_long(
  source_data,
  values_field,
  values_filter,
  value_type,
  cycles,
  time_unit,
  metadata,
  name_field = "display_name"
)
}
\description{
Extract Values as Long Format
}
\keyword{internal}
