% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_output.R
\name{extract_values_wide}
\alias{extract_values_wide}
\title{Extract Values as Wide Format}
\usage{
extract_values_wide(
  source_data,
  values_field,
  values_filter,
  value_type,
  cycles,
  time_unit,
  metadata,
  name_field = "display_name"
)
}
\description{
Extract Values as Wide Format
}
\keyword{internal}
