% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extract_var_from_ns}
\alias{extract_var_from_ns}
\title{Extract a variable value from a namespace}
\usage{
extract_var_from_ns(ns, name)
}
\description{
Extract a variable value from a namespace
}
\keyword{internal}
