% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{flatten_trees}
\alias{flatten_trees}
\title{Flatten Nested Tree Structure to Flat Tibble}
\usage{
flatten_trees(trees_list)
}
\description{
Flatten Nested Tree Structure to Flat Tibble
}
\keyword{internal}
