% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_ce_utils.R
\name{format_icer_label}
\alias{format_icer_label}
\title{Format ICER Value for Display}
\usage{
format_icer_label(
  icer_value,
  asterisk = "",
  decimals = 0,
  locale = NULL,
  abbreviate_val = FALSE
)
}
\arguments{
\item{icer_value}{Numeric ICER value}

\item{asterisk}{Character asterisk string to append}

\item{decimals}{Number of decimal places}

\item{locale}{Optional locale}

\item{abbreviate_val}{Logical. Use abbreviated values?}
}
\value{
Formatted string
}
\description{
Format ICER Value for Display
}
\keyword{internal}
