% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{format_multivariate_sampling_yaml}
\alias{format_multivariate_sampling_yaml}
\title{Format Multivariate Sampling for YAML}
\usage{
format_multivariate_sampling_yaml(mv_list)
}
\description{
Format Multivariate Sampling for YAML
}
\keyword{internal}
