% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{format_override_categories_yaml}
\alias{format_override_categories_yaml}
\title{Format Override Categories for YAML}
\usage{
format_override_categories_yaml(override_categories)
}
\description{
Format Override Categories for YAML
}
\keyword{internal}
