% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{format_threshold_yaml}
\alias{format_threshold_yaml}
\title{Format Threshold Analyses for YAML}
\usage{
format_threshold_yaml(threshold_list)
}
\description{
Format Threshold Analyses for YAML
}
\keyword{internal}
