% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_distributions.R
\name{gamma}
\alias{gamma}
\title{Gamma distribution}
\usage{
gamma(mean = NULL, sd = NULL, shape = NULL, scale = NULL)
}
\arguments{
\item{mean}{Mean of parameter (alternative parameterization)}

\item{sd}{Standard deviation of parameter (alternative parameterization)}

\item{shape}{Shape parameter (traditional parameterization)}

\item{scale}{Scale parameter (traditional parameterization)}
}
\description{
The `gamma` function can be used to sample a gamma-distributed
parameter, commonly used for costs or positive continuous parameters.
Can be specified using either (mean, sd) or (shape, scale) parameterization.
}
