% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_plots.R
\name{generate_ce_footnote}
\alias{generate_ce_footnote}
\title{Generate Footnote for Flipped Base Case}
\usage{
generate_ce_footnote(comparator_name, intervention_name)
}
\arguments{
\item{comparator_name}{Character display name of comparator}

\item{intervention_name}{Character display name of intervention}
}
\value{
Character footnote text
}
\description{
Internal helper to generate footnote text for flipped facets.
}
\keyword{internal}
