% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_codegen.R
\name{generate_decision_tree_code}
\alias{generate_decision_tree_code}
\title{Generate Decision Tree Code}
\usage{
generate_decision_tree_code(decision_tree)
}
\description{
Generate Decision Tree Code
}
\keyword{internal}
