% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_ce_utils.R
\name{generate_direction_change_footnote}
\alias{generate_direction_change_footnote}
\title{Generate Footnote for Direction Change}
\usage{
generate_direction_change_footnote(
  asterisk,
  more_costly_name,
  less_costly_name
)
}
\arguments{
\item{asterisk}{Character asterisk string}

\item{more_costly_name}{Character display name of more costly/effective strategy}

\item{less_costly_name}{Character display name of other strategy}
}
\value{
Character footnote text
}
\description{
Generate Footnote for Direction Change
}
\keyword{internal}
