% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_codegen.R
\name{generate_documentation_code}
\alias{generate_documentation_code}
\title{Generate Documentation Code}
\usage{
generate_documentation_code(documentation)
}
\description{
Generate Documentation Code
}
\keyword{internal}
