% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_ce_utils.R
\name{generate_flipped_bar_footnote}
\alias{generate_flipped_bar_footnote}
\title{Generate Footnote for Flipped Bar}
\usage{
generate_flipped_bar_footnote(asterisk, comparator_name, intervention_name)
}
\arguments{
\item{asterisk}{Character asterisk string}

\item{comparator_name}{Character comparator display name}

\item{intervention_name}{Character intervention display name}
}
\value{
Character footnote text
}
\description{
Generate Footnote for Flipped Bar
}
\keyword{internal}
