% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_ce_utils.R
\name{generate_identical_footnote}
\alias{generate_identical_footnote}
\title{Generate Footnote for Identical Outcomes}
\usage{
generate_identical_footnote(asterisk, intervention_name, comparator_name)
}
\arguments{
\item{asterisk}{Character asterisk string}

\item{intervention_name}{Character intervention display name}

\item{comparator_name}{Character comparator display name}
}
\value{
Character footnote text
}
\description{
Generate Footnote for Identical Outcomes
}
\keyword{internal}
