% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_codegen.R
\name{generate_transitions_code}
\alias{generate_transitions_code}
\title{Generate Transitions Code}
\usage{
generate_transitions_code(transitions, model_type = "markov")
}
\description{
Generate Transitions Code
}
\keyword{internal}
