% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twsa.R
\name{generate_twsa_values}
\alias{generate_twsa_values}
\title{Generate Values for TWSA Parameter}
\usage{
generate_twsa_values(
  param,
  namespace = NULL,
  settings = NULL,
  include_base_case = TRUE
)
}
\arguments{
\item{param}{Parameter specification from TWSA}

\item{namespace}{Evaluation namespace with variable values (for bc context)}

\item{settings}{Model settings (for setting base values)}

\item{include_base_case}{Logical. Ensure base case value is in the grid? (default: TRUE)}
}
\value{
Numeric vector of values
}
\description{
Evaluates the range specification for a TWSA parameter and returns a vector
of values to use.
}
\keyword{internal}
