% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{get_accumulated_errors}
\alias{get_accumulated_errors}
\title{Get Accumulated heRo Errors (Primarily for Testing)}
\usage{
get_accumulated_errors()
}
\value{
A list of accumulated error objects.
}
\description{
Returns the current internal list of accumulated errors without clearing it.
}
