% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{get_documentation}
\alias{get_documentation}
\title{Get Documentation}
\usage{
get_documentation(model)
}
\arguments{
\item{model}{An oq_model object}
}
\value{
The documentation string, or NULL if not set
}
\description{
Retrieve the markdown documentation string from the model.
}
