% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_dsa_parameters}
\alias{get_dsa_parameters}
\title{Get DSA Parameters}
\usage{
get_dsa_parameters(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A dsa_parameters object (list with class "dsa_parameters").
}
\description{
Get DSA Parameters
}
