% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_global_variables}
\alias{get_global_variables}
\title{Get Global Variables}
\usage{
get_global_variables(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A tibble of global variables.
}
\description{
Returns variables where strategy and group are both empty or NA.
}
