% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_model_summaries}
\alias{get_model_summaries}
\title{Get Model Summaries}
\usage{
get_model_summaries(model)
}
\arguments{
\item{model}{An oq_model object.}
}
\value{
A tibble of summary definitions.
}
\description{
Returns the summary definitions from the model object (not results).
Use \code{get_summaries()} from model results to get computed summaries.
}
