% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_model_value_names}
\alias{get_model_value_names}
\title{Get Model Value Names}
\usage{
get_model_value_names(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A character vector of unique value names.
}
\description{
Returns the unique value definition names from the model object.
}
