% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_model_values}
\alias{get_model_values}
\title{Get Model Values}
\usage{
get_model_values(model)
}
\arguments{
\item{model}{An oq_model object.}
}
\value{
A tibble of values.
}
\description{
Returns the value definitions from the model object (not results).
Use \code{get_values()} from model results to get computed values.
}
