% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_override_categories}
\alias{get_override_categories}
\title{Get Override Categories}
\usage{
get_override_categories(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A list of override categories.
}
\description{
Get Override Categories
}
