% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{get_results_locale}
\alias{get_results_locale}
\alias{get_results_number_locale}
\alias{get_results_currency_locale}
\title{Get Locale from Results Object}
\usage{
get_results_locale(results)

get_results_number_locale(results)

get_results_currency_locale(results)
}
\arguments{
\item{results}{A results object with metadata$settings$country}
}
\value{
A locale list (see \code{\link{get_locale}})
}
\description{
Extracts country from results metadata and returns the corresponding locale.
}
