% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_scenario_names}
\alias{get_scenario_names}
\title{Get Scenario Names}
\usage{
get_scenario_names(model)
}
\arguments{
\item{model}{An oq_model object.}
}
\value{
A character vector of scenario names.
}
\description{
Get Scenario Names
}
