% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_scripts}
\alias{get_scripts}
\title{Get Scripts}
\usage{
get_scripts(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A named list of scripts, each with \code{code} and \code{description}.
}
\description{
Get Scripts
}
