% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_settings}
\alias{get_settings}
\title{Get Settings}
\usage{
get_settings(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A list of model settings.
}
\description{
Get Settings
}
