% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_strategies}
\alias{get_strategies}
\title{Get Strategies}
\usage{
get_strategies(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A tibble of strategies.
}
\description{
Get Strategies
}
