% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_table_names}
\alias{get_table_names}
\title{Get Table Names}
\usage{
get_table_names(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A character vector of table names.
}
\description{
Get Table Names
}
