% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_tables}
\alias{get_tables}
\title{Get Tables}
\usage{
get_tables(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A named list of data frames.
}
\description{
Get Tables
}
