% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_plots.R
\name{get_threshold_input_label}
\alias{get_threshold_input_label}
\title{Get Threshold Input Label}
\usage{
get_threshold_input_label(analysis, metadata)
}
\arguments{
\item{analysis}{A single threshold analysis spec}

\item{metadata}{Model metadata}
}
\value{
Character string label
}
\description{
Returns a display name for the threshold variable.
}
\keyword{internal}
