% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_tree_names}
\alias{get_tree_names}
\title{Get Tree Names}
\usage{
get_tree_names(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A character vector of unique tree names.
}
\description{
Get Tree Names
}
