% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_trees}
\alias{get_trees}
\title{Get Trees}
\usage{
get_trees(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A data frame of trees, or NULL if none.
}
\description{
Get Trees
}
