% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_variable_names}
\alias{get_variable_names}
\title{Get Variable Names}
\usage{
get_variable_names(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A character vector of unique variable names.
}
\description{
Get Variable Names
}
