% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_variables}
\alias{get_variables}
\title{Get Variables}
\usage{
get_variables(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A tibble of variables.
}
\description{
Get Variables
}
