% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{is_oq_error}
\alias{is_oq_error}
\title{Check if Object is an openqaly Error}
\usage{
is_oq_error(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical; `TRUE` if `x` is an `oq_error`, `FALSE` otherwise.
}
\description{
Tests whether an object inherits from the `oq_error` class.
}
