% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{lf_to_arr}
\alias{lf_to_arr}
\title{Convert a longform table to a multi-dimensional array}
\usage{
lf_to_arr(df, dimcols, value)
}
\arguments{
\item{df}{A data frame in long format}

\item{dimcols}{Character vector of column names to use as dimensions}

\item{value}{Name of the column containing values}
}
\value{
A multi-dimensional array
}
\description{
Convert a longform table to a multi-dimensional array
}
\keyword{internal}
