% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.r
\name{lf_to_tmat}
\alias{lf_to_tmat}
\title{Convert Longform Transitions Table to Transition Matrix}
\usage{
lf_to_tmat(df)
}
\arguments{
\item{df}{A data frame with transition data}
}
\value{
A transition matrix array
}
\description{
Convert Longform Transitions Table to Transition Matrix
}
\keyword{internal}
