% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{model_builder}
\alias{model_builder}
\title{Model Builder Functions}
\description{
Functions for building heRomod2 models programmatically using a fluent API
with non-standard evaluation (NSE) for formula expressions.
}
