% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_codegen.R
\name{model_to_r_code}
\alias{model_to_r_code}
\title{Convert Model to R Code}
\usage{
model_to_r_code(model, file = NULL)
}
\arguments{
\item{model}{A oq_model object}

\item{file}{Optional file path to write the code to}
}
\value{
A character vector containing the R code lines
}
\description{
Generate executable R code that recreates a openqaly model.
}
\examples{
\dontrun{
# Generate R code as string
code <- model_to_r_code(model)

# Write R code to file
model_to_r_code(model, "my_model.R")
}
}
