% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_distributions.R
\name{mvnormal}
\alias{mvnormal}
\title{Multivariate Normal distribution}
\usage{
mvnormal(mean, cov)
}
\arguments{
\item{mean}{Mean vector (length k for k variables)}

\item{cov}{Covariance matrix (k x k)}
}
\value{
Function that takes n (number of samples) and returns n x k matrix
}
\description{
The `mvnormal` function can be used to sample correlated normal random variables.
Useful when parameters are known to be correlated (e.g., cost and effectiveness).
}
\examples{
\dontrun{
cov_mat <- matrix(c(0.01, 5, 5, 10000), nrow = 2)
dist_fn <- mvnormal(mean = c(0.5, 1000), cov = cov_mat)
samples <- dist_fn(1000)  # 1000 x 2 matrix
}
}
