% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{nest_threshold_analysis}
\alias{nest_threshold_analysis}
\title{Nest Threshold Analysis}
\usage{
nest_threshold_analysis(flat)
}
\arguments{
\item{flat}{A flat threshold analysis list}
}
\value{
A nested list with condition sub-list
}
\description{
Converts a flat threshold analysis to nested format (from Excel deserialization).
}
\keyword{internal}
