% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{nest_trees}
\alias{nest_trees}
\title{Convert Flat Tree Tibble to Nested Structure for YAML}
\usage{
nest_trees(trees_tbl)
}
\description{
Convert Flat Tree Tibble to Nested Structure for YAML
}
\keyword{internal}
