% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{notify_active_overrides}
\alias{notify_active_overrides}
\title{Emit Override Notification Message}
\usage{
notify_active_overrides(model)
}
\arguments{
\item{model}{Model object with override_categories}
}
\description{
Prints an informational message about active overrides being applied.
}
\keyword{internal}
