% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_tables.R
\name{outcomes_table_flextable_old}
\alias{outcomes_table_flextable_old}
\title{OLD FLEXTABLE IMPLEMENTATION - KEPT FOR REFERENCE, DELETE AFTER TESTING}
\usage{
outcomes_table_flextable_old(
  results,
  summary_name,
  group = "aggregated",
  strategies = NULL,
  show_total = TRUE,
  decimals = 2,
  discounted = FALSE,
  value_name_field = "display_name",
  strategy_name_field = "display_name",
  group_name_field = "display_name"
)
}
\description{
OLD FLEXTABLE IMPLEMENTATION - KEPT FOR REFERENCE, DELETE AFTER TESTING
}
\keyword{internal}
