% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_model.R
\name{parse_decision_tree_model}
\alias{parse_decision_tree_model}
\title{Parse Decision Tree Model}
\usage{
parse_decision_tree_model(model)
}
\arguments{
\item{model}{The model object}
}
\value{
The model with class set to 'decision_tree'
}
\description{
Sets up a standalone decision tree model for execution.
}
\keyword{internal}
