% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{parse_embedded_csv}
\alias{parse_embedded_csv}
\title{Parse Embedded CSV String into Data Frame}
\usage{
parse_embedded_csv(csv_string)
}
\description{
Parse Embedded CSV String into Data Frame
}
\keyword{internal}
