% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm.R
\name{parse_psm_custom}
\alias{parse_psm_custom}
\title{Parse Custom PSM Model}
\usage{
parse_psm_custom(model)
}
\arguments{
\item{model}{A heRomodel object}
}
\value{
The model with psm_custom class
}
\description{
Validates the structure of a Custom PSM model. Custom PSM models allow
arbitrary numbers of states with direct probability formulas.
}
\keyword{internal}
