% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.r
\name{parse_summaries}
\alias{parse_summaries}
\title{Parse and validate summaries}
\usage{
parse_summaries(summaries, model_values)
}
\arguments{
\item{summaries}{A dataframe containing summary definitions}

\item{model_values}{Vector of value names available in the model}
}
\value{
A validated summaries dataframe with parsed values
}
\description{
Parse and validate summaries
}
