% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{parse_tree_vars}
\alias{parse_tree_vars}
\title{Define Decision Trees}
\usage{
parse_tree_vars(trees)
}
\arguments{
\item{trees}{A dataframe containing decision tree specifications with columns
name, node, parent, formula, and tags}
}
\value{
A tibble of variable definitions for the decision trees
}
\description{
Parses a decision tree specification dataframe into variable definitions.
}
\keyword{internal}
