% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{parse_yaml_multivariate_sampling}
\alias{parse_yaml_multivariate_sampling}
\title{Parse YAML Multivariate Sampling}
\usage{
parse_yaml_multivariate_sampling(mv_list)
}
\description{
Parse YAML Multivariate Sampling
}
\keyword{internal}
