% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{parse_yaml_override_categories}
\alias{parse_yaml_override_categories}
\title{Parse YAML Override Categories}
\usage{
parse_yaml_override_categories(oc_list)
}
\description{
Parse YAML Override Categories
}
\keyword{internal}
