% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{parse_yaml_threshold}
\alias{parse_yaml_threshold}
\title{Parse YAML Threshold Analyses}
\usage{
parse_yaml_threshold(threshold_list)
}
\description{
Parse YAML Threshold Analyses
}
\keyword{internal}
