% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_data.R
\name{prepare_dsa_tornado_data}
\alias{prepare_dsa_tornado_data}
\title{Compatibility alias for older internal name}
\usage{
prepare_dsa_tornado_data(
  results,
  summary_name,
  groups,
  strategies = NULL,
  interventions = NULL,
  comparators = NULL,
  discounted = TRUE,
  show_parameter_values = TRUE,
  locale = NULL,
  value_type = "all"
)
}
\description{
Compatibility alias for older internal name
}
\keyword{internal}
