% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{print.dsa_parameters}
\alias{print.dsa_parameters}
\title{Print DSA Parameters}
\usage{
\method{print}{dsa_parameters}(x, ...)
}
\arguments{
\item{x}{A dsa_parameters object (list)}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for dsa_parameters objects
}
