% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{print_scenarios}
\alias{print_scenarios}
\title{Print Scenarios}
\usage{
print_scenarios(model)
}
\arguments{
\item{model}{An oq_model_builder object with scenarios}
}
\value{
Invisible model
}
\description{
Print method for displaying scenarios defined in a model
}
\keyword{internal}
