% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{print_twsa}
\alias{print_twsa}
\title{Print TWSA Analyses}
\usage{
print_twsa(model)
}
\arguments{
\item{model}{An oq_model_builder object with TWSA analyses}
}
\value{
Invisible model
}
\description{
Print method for displaying two-way sensitivity analyses defined in a model
}
\keyword{internal}
