% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_custom.R
\name{prune_namespace_cycle_zero}
\alias{prune_namespace_cycle_zero}
\title{Prune Cycle Zero from Namespace}
\usage{
prune_namespace_cycle_zero(namespace)
}
\arguments{
\item{namespace}{A namespace object containing cycle 0}
}
\value{
A namespace object with cycle 0 removed from df
}
\description{
Removes the first row (cycle 0) from the namespace data frame.
Used to prepare namespace for values calculation which only uses cycles 1:n_cycles.
}
\keyword{internal}
