% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_model_json}
\alias{read_model_json}
\title{Read a model from JSON.}
\usage{
read_model_json(file = NULL, text = NULL)
}
\arguments{
\item{file}{Path to a JSON file containing the model.}

\item{text}{A string containing the model in JSON format.}
}
\value{
An oq_model object
}
\description{
Supply exactly one of \code{file} or \code{text}.
}
