% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_custom_psm_transition}
\alias{remove_custom_psm_transition}
\title{Remove a Custom PSM Transition}
\usage{
remove_custom_psm_transition(model, state)
}
\arguments{
\item{model}{An oq_model_builder object (must be type "custom_psm")}

\item{state}{Character string for the state name}
}
\value{
The modified model object
}
\description{
Remove an existing Custom PSM state probability.
}
