% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_decision_tree}
\alias{remove_decision_tree}
\title{Remove Decision Tree Configuration}
\usage{
remove_decision_tree(model)
}
\arguments{
\item{model}{An oq_model_builder object}
}
\value{
The modified model object
}
\description{
Remove the decision tree configuration from the model.
}
