% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_dsa_setting}
\alias{remove_dsa_setting}
\title{Remove a DSA Setting Specification}
\usage{
remove_dsa_setting(model, setting)
}
\arguments{
\item{model}{A oq_model_builder object}

\item{setting}{Character string naming the setting to remove}
}
\value{
The modified model object
}
\description{
Remove an existing DSA setting specification from the model.
}
\examples{
\dontrun{
model <- define_model("markov") |>
  set_settings(discount_cost = 3) |>
  add_dsa_setting("discount_cost", low = 0, high = 5) |>
  remove_dsa_setting("discount_cost")
}
}
