% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_lockfile}
\alias{remove_lockfile}
\title{Remove Lockfile from Model}
\usage{
remove_lockfile(model)
}
\arguments{
\item{model}{An openqaly model object.}
}
\value{
The model with \code{lockfile} field removed.
}
\description{
Remove Lockfile from Model
}
