% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_multivariate_sampling}
\alias{remove_multivariate_sampling}
\title{Remove a Multivariate Sampling Specification}
\usage{
remove_multivariate_sampling(model, name)
}
\arguments{
\item{model}{A oq_model_builder object}

\item{name}{Character string identifying the specification to remove}
}
\value{
The modified model object
}
\description{
Remove an existing multivariate sampling specification from the model.
}
\examples{
\dontrun{
model <- model |>
  remove_multivariate_sampling("transition_probs")
}
}
