% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_override_category}
\alias{remove_override_category}
\title{Remove an Override Category}
\usage{
remove_override_category(model, name)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{name}{Character string identifying the category to remove}
}
\value{
The modified model object
}
\description{
Remove an existing override category and all its overrides.
}
