% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_psm_transition}
\alias{remove_psm_transition}
\title{Remove a PSM Transition}
\usage{
remove_psm_transition(model, endpoint)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{endpoint}{Character string for the endpoint}
}
\value{
The modified model object
}
\description{
Remove an existing PSM transition.
}
