% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_scenario_setting}
\alias{remove_scenario_setting}
\title{Remove a Scenario Setting Override}
\usage{
remove_scenario_setting(model, scenario, setting)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{scenario}{Name of the scenario containing the override}

\item{setting}{Name of the setting to remove}
}
\value{
The modified model object
}
\description{
Remove an existing setting override from a scenario.
}
\examples{
\dontrun{
model <- define_model("markov") |>
  set_settings(timeframe = 20) |>
  add_scenario("Extended") |>
  add_scenario_setting("Extended", "timeframe", 30) |>
  remove_scenario_setting("Extended", "timeframe")
}
}
