% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_script}
\alias{remove_script}
\title{Remove a Script from the Model}
\usage{
remove_script(model, name)
}
\arguments{
\item{model}{A oq_model_builder object}

\item{name}{Character string identifying the script to remove}
}
\value{
The modified model object
}
\description{
Remove an existing script from the model by name.
}
