% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_table}
\alias{remove_table}
\title{Remove a Table from the Model}
\usage{
remove_table(model, name)
}
\arguments{
\item{model}{A oq_model_builder object}

\item{name}{Character string identifying the table to remove}
}
\value{
The modified model object
}
\description{
Remove an existing table from the model by name.
}
