% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_transition}
\alias{remove_transition}
\title{Remove a Markov Transition}
\usage{
remove_transition(model, from_state, to_state)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{from_state}{Character string specifying the source state}

\item{to_state}{Character string specifying the destination state}
}
\value{
The modified model object
}
\description{
Remove an existing Markov transition.
}
