% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_tree}
\alias{remove_tree}
\title{Remove an Entire Decision Tree}
\usage{
remove_tree(model, tree_name)
}
\arguments{
\item{model}{An oq_model object}

\item{tree_name}{Character string for the tree name}
}
\value{
The modified model object
}
\description{
Remove all nodes of a tree from the model.
}
