% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_tree_node}
\alias{remove_tree_node}
\title{Remove a Decision Tree Node}
\usage{
remove_tree_node(model, tree_name, node)
}
\arguments{
\item{model}{An oq_model object}

\item{tree_name}{Character string for the tree name}

\item{node}{Character string for the node name}
}
\value{
The modified model object
}
\description{
Remove a tree node and all its descendants from the model.
}
