% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_twsa_setting}
\alias{remove_twsa_setting}
\title{Remove a TWSA Setting Parameter}
\usage{
remove_twsa_setting(model, twsa_name, setting)
}
\arguments{
\item{model}{An oq_model_builder object}

\item{twsa_name}{Name of the TWSA analysis}

\item{setting}{Name of the setting to remove}
}
\value{
The modified model object
}
\description{
Remove an existing setting parameter from a TWSA analysis.
}
\examples{
\dontrun{
model <- define_model("markov") |>
  add_twsa("Test") |>
  add_twsa_setting("Test", "discount_cost", type = "range",
    min = 0, max = 5, steps = 3) |>
  remove_twsa_setting("Test", "discount_cost")
}
}
