% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_twsa_variable}
\alias{remove_twsa_variable}
\title{Remove a TWSA Variable Parameter}
\usage{
remove_twsa_variable(model, twsa_name, variable, strategy = "", group = "")
}
\arguments{
\item{model}{An oq_model_builder object}

\item{twsa_name}{Name of the TWSA analysis}

\item{variable}{Name of the variable to remove}

\item{strategy}{Strategy targeting used to identify the parameter (default: "")}

\item{group}{Group targeting used to identify the parameter (default: "")}
}
\value{
The modified model object
}
\description{
Remove an existing variable parameter from a TWSA analysis.
}
\examples{
\dontrun{
model <- define_model("markov") |>
  add_twsa("Test") |>
  add_twsa_variable("Test", "cost", type = "range",
    min = 500, max = 1500, steps = 5) |>
  remove_twsa_variable("Test", "cost")
}
}
