% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{remove_vbp}
\alias{remove_vbp}
\title{Remove VBP Configuration}
\usage{
remove_vbp(model)
}
\arguments{
\item{model}{An oq_model_builder object}
}
\value{
The modified model object
}
\description{
Remove the value-based pricing (VBP) configuration from the model.
}
