% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_table_spec.R, R/table_renderers.R
\name{render_kable_simple}
\alias{render_kable_simple}
\title{Render Simple Table Spec as Kable}
\usage{
render_kable_simple(spec)

render_kable_simple(spec)
}
\arguments{
\item{spec}{A simple table specification}
}
\value{
A kableExtra HTML table

A kableExtra HTML table
}
\description{
Clean renderer that converts spec to kableExtra HTML table.
}
\keyword{internal}
