% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_codegen.R
\name{resolve_val}
\alias{resolve_val}
\title{Convert oq_formula to character, pass through other types}
\usage{
resolve_val(x)
}
\description{
Convert oq_formula to character, pass through other types
}
\keyword{internal}
