% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_codegen.R
\name{rv_values}
\alias{rv_values}
\title{Render custom values vector, handling oq_formula}
\usage{
rv_values(vals, is_var)
}
\description{
Render custom values vector, handling oq_formula
}
\keyword{internal}
