% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_vbp.R
\name{scenario_vbp}
\alias{scenario_vbp}
\title{Scenario Analysis with Value-Based Pricing Add-on}
\description{
Functions for combining Scenario Analysis with Value-Based Pricing (VBP) analysis.
For each scenario, VBP sub-simulations are run to calculate the VBP relationship
for each scenario.
}
