% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{set_override_categories}
\alias{set_override_categories}
\title{Set Override Categories}
\usage{
set_override_categories(model, categories)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}

\item{categories}{A list of override categories.}
}
\value{
The modified model object.
}
\description{
Replaces override categories wholesale on a model object.
}
