% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbp.R
\name{test_vbp_nmb_consistency}
\alias{test_vbp_nmb_consistency}
\title{Test VBP NMB Consistency}
\usage{
test_vbp_nmb_consistency()
}
\value{
TRUE if all tests pass
}
\description{
Tests that VBP calculations are correct by verifying that at the calculated
VBP price, the Net Monetary Benefit (NMB) equals zero (i.e., the intervention
is exactly cost-effective at the given WTP threshold).
}
\examples{
\dontrun{
test_vbp_nmb_consistency()
}

}
