% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold_optimizer}
\alias{threshold_optimizer}
\title{Optimizer Fallback for Threshold}
\usage{
threshold_optimizer(solver_callback, analysis)
}
\description{
Optimizer Fallback for Threshold
}
\keyword{internal}
