% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold_root_finder}
\alias{threshold_root_finder}
\title{Root Finder for Threshold}
\usage{
threshold_root_finder(solver_callback, analysis)
}
\description{
Root Finder for Threshold
}
\keyword{internal}
