% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{tibble_to_yaml_list}
\alias{tibble_to_yaml_list}
\title{Convert Tibble to YAML-friendly List}
\usage{
tibble_to_yaml_list(tbl)
}
\description{
Convert Tibble to YAML-friendly List
}
\keyword{internal}
