% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_distributions.R
\name{triangular}
\alias{triangular}
\title{Triangular distribution}
\usage{
triangular(min, mode, max)
}
\arguments{
\item{min}{Minimum value}

\item{mode}{Mode (peak) value}

\item{max}{Maximum value}
}
\description{
The `triangular` function can be used to sample a triangular-distributed
parameter, useful when only min, mode, and max are known (e.g., from expert opinion).
}
