% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_distributions.R
\name{uniform}
\alias{uniform}
\title{Uniform distribution}
\usage{
uniform(min, max)
}
\arguments{
\item{min}{Minimum value}

\item{max}{Maximum value}
}
\description{
The `uniform` function can be used to sample a uniformly-distributed
parameter between specified bounds.
}
