% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_settings}
\alias{validate_settings}
\title{Normalize and Validate Model Structure}
\usage{
validate_settings(settings, model_type)
}
\arguments{
\item{settings}{A list of model settings}

\item{model_type}{The canonical model type string}
}
\description{
Validate model settings
}
\keyword{internal}
