% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{validate_threshold_condition}
\alias{validate_threshold_condition}
\title{Validate Threshold Condition Fields}
\usage{
validate_threshold_condition(condition)
}
\arguments{
\item{condition}{A threshold condition list}
}
\description{
Validate Threshold Condition Fields
}
\keyword{internal}
