% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{validate_tree_name_collisions}
\alias{validate_tree_name_collisions}
\title{Validate tree name collisions with other model components}
\usage{
validate_tree_name_collisions(tree_names, model)
}
\arguments{
\item{tree_names}{Character vector of tree names to check}

\item{model}{The model object}
}
\description{
Validate tree name collisions with other model components
}
\keyword{internal}
