% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_converters.R
\name{write_model_excel}
\alias{write_model_excel}
\title{Write Model as Excel}
\usage{
write_model_excel(model, path)
}
\arguments{
\item{model}{A oq_model object}

\item{path}{Folder path}
}
\description{
Internal function to write a model to Excel folder structure.
}
\keyword{internal}
