% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_model_json}
\alias{write_model_json}
\title{Write Model to JSON}
\usage{
write_model_json(model)
}
\arguments{
\item{model}{A heRomodel object to be converted to JSON.}
}
\value{
A JSON string representing the model.
}
\description{
Takes a heRomodel object (typically loaded from Excel) and converts it
to a JSON string format.
}
