% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm.R
\name{convert_cycles_to_time_unit}
\alias{convert_cycles_to_time_unit}
\title{Convert Cycles to Time Unit}
\usage{
convert_cycles_to_time_unit(cycles, time_unit, namespace)
}
\arguments{
\item{cycles}{Vector of cycle numbers}

\item{time_unit}{The target time unit (days/weeks/months/years)}

\item{namespace}{The namespace containing time variables}
}
\value{
Vector of times in the specified unit
}
\description{
Converts cycle numbers to the appropriate time unit for survival distributions.
}
\keyword{internal}
