% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert_json_dataframes}
\alias{convert_json_dataframes}
\title{Convert JSON data frames}
\usage{
convert_json_dataframes(model, values_spec = NULL)
}
\arguments{
\item{model}{The model list parsed from JSON}

\item{values_spec}{Optional values specification list}
}
\value{
The model with properly formatted data frames
}
\description{
Ensures that data frames in the model are properly formatted
}
