% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa_plots.R
\name{generate_direction_change_footnote}
\alias{generate_direction_change_footnote}
\title{Generate Footnote for Direction Change}
\usage{
generate_direction_change_footnote(
  asterisk,
  more_costly_name,
  less_costly_name
)
}
\arguments{
\item{asterisk}{Character asterisk string (e.g., "*", "**", "***")}

\item{more_costly_name}{Character display name of the strategy that is more costly/effective}

\item{less_costly_name}{Character display name of the other strategy}
}
\value{
Character footnote text
}
\description{
Internal helper to generate footnote text for direction change errors.
}
\keyword{internal}
