% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_model_type}
\alias{get_model_type}
\title{Get Model Type}
\usage{
get_model_type(model)
}
\arguments{
\item{model}{An oq_model or oq_model_builder object.}
}
\value{
A string indicating the model type (e.g., "markov", "psm").
}
\description{
Get Model Type
}
