% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api.R
\name{get_script_names}
\alias{get_script_names}
\title{Get Script Names}
\usage{
get_script_names(model)
}
\arguments{
\item{model}{An oq_model object.}
}
\value{
A character vector of script names.
}
\description{
Get Script Names
}
