% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{hero_error_checkpoint}
\alias{hero_error_checkpoint}
\title{Check for and report errors at a checkpoint (INTERNAL).}
\usage{
hero_error_checkpoint()
}
\value{
Invisibly returns NULL.
}
\description{
Called by package functions at specific points.
This checks the internal error list (.hero_error_env$errors).
If non-dependency errors exist, they are formatted and thrown via `format_and_throw_errors`.
Regardless of whether errors were thrown, the internal error list is cleared.
}
