% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_table_prepare.R
\name{prepare_simple_outcomes_table}
\alias{prepare_simple_outcomes_table}
\title{Prepare Simple Outcomes Table Specification}
\usage{
prepare_simple_outcomes_table(
  results,
  summary_name,
  group = "aggregated",
  strategies = NULL,
  show_total = TRUE,
  decimals = 2,
  discounted = FALSE,
  value_name_field = "display_name",
  font_size = 11
)
}
\arguments{
\item{results}{openqaly results object}

\item{summary_name}{Name of summary to display}

\item{group}{Group selection}

\item{strategies}{Character vector of strategies to include}

\item{show_total}{Show TOTAL row}

\item{decimals}{Number of decimal places}

\item{discounted}{Use discounted values}

\item{value_name_field}{Field to use for value names}

\item{font_size}{Font size in points}
}
\value{
A simple table specification
}
\description{
Prepare Simple Outcomes Table Specification
}
\keyword{internal}
