% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.r
\name{print.icer}
\alias{print.icer}
\title{Printing for ICER vectors
Formats values as:
  +number = more costly/more effective (ICER)
  -number = less costly/less effective (prints as opposite with asterisk)
  0       = "Dominant"
  Inf     = "Dominated"
  NaN     = "Equivalent"
  NA      = "" (blank - reference strategy with no comparison)}
\usage{
\method{print}{icer}(x, digits = 3, big.mark = ",", ...)
}
\description{
Printing for ICER vectors
Formats values as:
  +number = more costly/more effective (ICER)
  -number = less costly/less effective (prints as opposite with asterisk)
  0       = "Dominant"
  Inf     = "Dominated"
  NaN     = "Equivalent"
  NA      = "" (blank - reference strategy with no comparison)
}
