% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_model_json}
\alias{read_model_json}
\title{Takes a JSON string and parses it into a heRomodel object.}
\usage{
read_model_json(json_string)
}
\arguments{
\item{json_string}{A string containing the model in JSON format.}
}
\value{
A heRomodel object
}
\description{
Takes a JSON string and parses it into a heRomodel object.
}
